requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --output-file requirements.txt --no-hashes
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via tqdm
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.19.1
    # via torch
flatbuffers==25.2.10
    # via onnxruntime
fonttools==4.59.1
    # via matplotlib
fsspec==2025.7.0
    # via torch
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via requests
jinja2==3.1.6
    # via torch
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.5
    # via
    #   background-removal
    #   ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.2.6
    # via
    #   contourpy
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   pandas
    #   pycocotools
    #   scipy
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
nvidia-cublas-cu12==12.8.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cuda-runtime-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cudnn-cu12==9.10.2.21 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufft-cu12==11.3.3.83 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cufile-cu12==1.13.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-curand-cu12==10.3.9.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusolver-cu12==11.7.3.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-cusparse-cu12==12.5.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nccl-cu12==2.27.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
nvidia-nvjitlink-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
onnx==1.18.0
    # via background-removal
onnxruntime==1.22.1
    # via background-removal
opencv-python==4.12.0.88
    # via
    #   background-removal
    #   ultralytics
packaging==25.0
    # via
    #   matplotlib
    #   onnxruntime
pandas==2.3.1
    # via ultralytics
pillow==11.3.0
    # via
    #   matplotlib
    #   torchvision
    #   ultralytics
protobuf==6.32.0
    # via
    #   onnx
    #   onnxruntime
psutil==7.0.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pycocotools==2.0.10
    # via background-removal
pyparsing==3.2.3
    # via matplotlib
pyreadline3==3.5.4 ; sys_platform == 'win32'
    # via humanfriendly
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via ultralytics
requests==2.32.4
    # via ultralytics
scipy==1.16.1
    # via ultralytics
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
    # via background-removal
setuptools==80.9.0
    # via
    #   torch
    #   triton
six==1.17.0
    # via python-dateutil
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
torch==2.8.0
    # via
    #   background-removal
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.23.0
    # via
    #   background-removal
    #   ultralytics
tqdm==4.67.1
    # via ultralytics
triton==3.4.0 ; platform_machine == 'x86_64' and sys_platform == 'linux'
    # via torch
typing-extensions==4.14.1
    # via
    #   onnx
    #   torch
tzdata==2025.2
    # via pandas
ultralytics==8.3.179
    # via background-removal
ultralytics-thop==2.0.15
    # via ultralytics
urllib3==2.5.0
    # via requests
